backend/requirements.txt
Flask>=3.0
SQLAlchemy>=1.4
neo4j>=5.0
python-dotenv
requests
pytest